scaffold-api/requirements/prod.txt
gunicorn
Flask>=3.0.0
Flask-Migrate
Flask-Mail
Flask-Moment
Flask-SQLAlchemy
SQLAlchemy-Continuum
flask-restx
flask-marshmallow==1.2.1
flask-jwt-oidc
python-dotenv
psycopg2-binary
marshmallow-sqlalchemy==1.0.0
marshmallow_enum
jsonschema
requests
itsdangerous
werkzeug
sqlalchemy-utils
Flask-Caching
sqlalchemy
secure
python-dotenv
requests
flask_cors
pyhumps
importlib-resources